Refactor logLines to SplitByNewline, deduplicate
