+++ internal/testutil/stringutil/stringutil.go
@@ -1,11 +1,11 @@
 // Copyright 2024 the Pinniped contributors. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package testlogger
+package stringutil
 
 import "strings"
 
-func LogLines(logs string) []string {
+func SplitByNewline(logs string) []string {
 	if len(logs) == 0 {
 		return nil
 	}
